solved problems with global statics
