+++ core/src/soots/arthropods.rs
@@ -1,5 +1,6 @@
 use nannou::prelude::*;
-use core::cfg_fetch;
+use crate::cfg_fetch;
+use std::path::Path;
 
 /// arthropods define enemy types
 #[derive(Debug, Clone, Copy)]
@@ -10,8 +11,10 @@ pub enum Arthropod {
 /// Picks up the texture for each type
 
 impl Arthropod {
-    fn toTexture(&self) -> wgpu::Texture {
-        cfg_fetch();
+
+    fn toTexture(&self, app: &App) -> wgpu::Texture {
+        let texture_path = Path::new(self.assets()).join(Path::new(&cfg_fetch(&self.toConfig())));
+        wgpu::Texture::from_path(app, texture_path).unwrap()
     }
 
     fn toConfig(&self) -> String {
@@ -20,5 +23,10 @@ impl Arthropod {
             _ => "none"
         }
 
+    }
+
+    fn assets(&self) -> &Path {
+        let asset_pth = cfg_fetch("assets.sprites");
+        &Path::new(&asset_pth)
     }
 }
